Written Analysis of Results
# Kickstarting with Excel

## Overview of Project
This project involved utilizing what we learned in Module 1 in Excel to help Louise analyze the data she has for kickstarters. Louise had a data file with information about her kickstarter campaigns, and we analyze the data through pivot tables, descriptive statistics, and visualizations.

### Purpose
The purpose of this project was to help Louise better understand and utilize her data. We specifically looked at the theater category. Since Louise had a fairly large data set, it was important for us to both organize and visualize the data in a way that it could be used for analysis. 

## Analysis and Challenges
The analysis included the use of pivot tables and charts to visualize the data. The two analyses we used for this project were: theater outcomes by launch date and outcomes based on goals. These data points came from the goal of each theater campaign, which is a monetary amount, and the date of the theater category campaign's launch. 

### Analysis of Outcomes Based on Launch Date
When reviewing the outcomes of the theater campaign by launch date, we looked at whether the campaign was successful, canceled, failed, or live. In the pivot table, we filtered out any live campaigns and looked solely at successful, failed, and cancelled campaigns. The chart helped visualize how these categories vary by month. You can immediately see that there was a spike in successful theater campaigns in May and June. Other than the spike in those two months, the data seems to be linear between successful and failed launches by month. There were almost 400 more successful campaigns than failed campaigns in the theater category and that was primarily driven by the May and June months. It appears that those are the best months to launch a theater campaign! December, on the other hand, had the least amount of successful campaigns, and almost the same amount of failed campaigns as successful ones.

### Analysis of Outcomes Based on Goals
The next data we analyzed was the theater campaigns based on the kickstarter's goal. We visualized the data using a range of goals in dollar amounts in increments in $5,000. For this analysis, we specifically looked at the subcategory of plays. While the majority of these campaigns had less than $5,000 goals, we visualized the data based on the percentage of success and percentage of failed by goal amount so the actual dollar amount did not impact our visualization. Plays with goals between $0 and $15,000 and between $35,000 and $45,000 had a greater rate of success than failure, while plays with goals between $20,000 and $35,000 and greater than $45,000 had a higher rate of failure than success.

### Challenges and Difficulties Encountered
The difficulties and challenges encountered with this data is there are many factors that can impact the success or failure of a campaign. We did not filter the data based on region or year, and these factors could impact the result of the campaign. There could also be a correlation between the number of campaigns running at a specific time. I think looking at the data by region would help us assist Louise better. 

## Results
The results of this project yielded information about what goals should be set for successful play campaigns in the theater category. Goals less than $15,000 and between $35,000 and $45,000 result in more successful campaigns. Additionally, launching these campaigns in May or June appears to be the best timeframe when looking at the month the campaign is launched.

## Conclusions
Two conclusions you can draw from the launch date analysis is that warmer months lead to more successful campaigns. The summer months, May, June, and July had the highest number of successful campaigns. Additionally, the colder months appear to be the worst time to launch a theater campaign. December had the lowest count of successful campaigns. 

One conclusion you can draw from the outcomes by goals analysis is that you do not need have extremely high goals to have a successful campaign. In fact, the highest goal amount categories resulted in the highest rate of failed campaigns. 

## Limitations and Possible Changes
The limitations of this data set are that there are other variables that go into a successful campaign than just the launch date and goal. Adding a filter for the country that a campaign is launched in could help the user look at the data by geographic region. Additionally, looking at this data by year would help the user visualize how many campaigns are launched over a specified period of time. Depending on the user's goal, you can visualize the data in a number of different ways to better develop hypotheses and strategies for the future.